# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "FrameInput": {
        "additionalProperties": false,
        "properties": {
          "prices": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Prices",
            "type": "object"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          },
          "trading_day": {
            "title": "Trading Day",
            "type": "string"
          }
        },
        "required": [
          "sequence",
          "trading_day",
          "prices"
        ],
        "title": "FrameInput",
        "type": "object"
      },
      "FrameResponse": {
        "properties": {
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "sequence": {
            "title": "Sequence",
            "type": "integer"
          }
        },
        "required": [
          "sequence",
          "duplicate"
        ],
        "title": "FrameResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrderInput": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "string"
          },
          "offset": {
            "enum": [
              "OPEN",
              "CLOSE_TODAY",
              "CLOSE_YESTERDAY"
            ],
            "title": "Offset",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "side": {
            "enum": [
              "LONG",
              "SHORT"
            ],
            "title": "Side",
            "type": "string"
          },
          "trading_day": {
            "title": "Trading Day",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "account_id",
          "instrument_id",
          "trading_day",
          "side",
          "offset",
          "quantity"
        ],
        "title": "OrderInput",
        "type": "object"
      },
      "OrderResponse": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "duplicate": {
            "title": "Duplicate",
            "type": "boolean"
          },
          "fee": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fee"
          },
          "fill_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fill Id"
          },
          "frame_sequence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Sequence"
          },
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "FILLED",
              "REJECTED"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "request_id",
          "account_id",
          "status",
          "reason",
          "frame_sequence",
          "fill_id",
          "price",
          "fee",
          "duplicate"
        ],
        "title": "OrderResponse",
        "type": "object"
      },
...
      "RiskEventResponse": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "detected_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Ms"
          },
          "event_id": {
            "title": "Event Id",
            "type": "integer"
          },
          "frame_sequence": {
            "title": "Frame Sequence",
            "type": "integer"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "occurred_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred Ms"
          },
          "product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Product Id"
          },
          "recovered": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recovered"
          },
          "trading_day": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trading Day"
          }
        },
        "required": [
          "event_id",
          "account_id",
          "frame_sequence",
          "kind",
          "product_id"
        ],
        "title": "RiskEventResponse",
        "type": "object"
      },
      "RiskResponse": {
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "available_funds": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Available Funds"
          },
          "blocking_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Blocking Reasons",
            "type": "array"
          },
          "circuit_broken": {
            "title": "Circuit Broken",
            "type": "boolean"
          },
          "equity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equity"
          },
          "floating_pnl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Floating Pnl"
          },
          "frame_sequence": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Frame Sequence"
          },
          "gross_exposure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gross Exposure"
          },
          "margin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Margin"
          },
          "market_ready": {
            "title": "Market Ready",
            "type": "boolean"
          },
          "opening_allowed": {
            "title": "Opening Allowed",
            "type": "boolean"
          },
          "restricted_products": {
            "items": {
              "type": "string"
            },
            "title": "Restricted Products",
            "type": "array"
          },
          "trading_day": {
            "format": "date",
            "title": "Trading Day",
            "type": "string"
          },
          "warning": {
            "title": "Warning",
            "type": "boolean"
          }
        },
        "required": [
          "account_id",
          "trading_day",
          "frame_sequence",
          "market_ready",
          "warning",
          "circuit_broken",
          "restricted_products",
          "opening_allowed",
          "blocking_reasons",
          "floating_pnl",
          "equity",
          "margin",
          "available_funds",
          "gross_exposure"
        ],
        "title": "RiskResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "CTA 期货风控 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/orders": {
      "get": {
        "operationId": "listOrders",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/OrderResponse"
                  },
                  "title": "Response Listorders",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Orders",
        "tags": [
          "trading"
        ]
      },
      "post": {
        "operationId": "submitOrder",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Order",
        "tags": [
          "trading"
        ]
      }
    },
    "/api/orders/{account_id}/{request_id}": {
      "get": {
        "operationId": "getOrder",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "request_id",
            "required": true,
            "schema": {
              "title": "Request Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Order",
        "tags": [
          "trading"
        ]
      }
    },
...
    "/api/risk": {
      "get": {
        "operationId": "listAccountRisk",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RiskResponse"
                  },
                  "title": "Response Listaccountrisk",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk",
        "tags": [
          "trading"
        ]
      }
    },
    "/api/risk-events": {
      "get": {
        "operationId": "listRiskEvents",
        "parameters": [
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RiskEventResponse"
                  },
                  "title": "Response Listriskevents",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Risk Events",
        "tags": [
          "trading"
        ]
      }
    },
...
    "/api/simulation/frames": {
      "post": {
        "operationId": "publishSimulationFrame",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FrameInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FrameResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Frame",
        "tags": [
          "trading"
        ]
      }
    },
...
  }
}